refactor(assistant): move scoped-approval-grants out of memory/ into approvals/
