+++ assistant/src/__tests__/scoped-approval-grants.test.ts
@@ -13,7 +13,7 @@ import {
   type CreateScopedApprovalGrantParams,
   expireScopedApprovalGrants,
   revokeScopedApprovalGrantsForContext,
-} from "../memory/scoped-approval-grants.js";
+} from "../approvals/scoped-approval-grants.js";
 import { getDb } from "../persistence/db-connection.js";
 import { initializeDb } from "../persistence/db-init.js";
 import { scopedApprovalGrants } from "../persistence/schema/index.js";
