+++ assistant/src/__tests__/scoped-grant-security-matrix.test.ts
@@ -36,7 +36,7 @@ mock.module("../util/logger.js", () => ({
 import {
   _internal,
   type CreateScopedApprovalGrantParams,
-} from "../memory/scoped-approval-grants.js";
+} from "../approvals/scoped-approval-grants.js";
 import { getDb } from "../persistence/db-connection.js";
 import { initializeDb } from "../persistence/db-init.js";
 import { scopedApprovalGrants } from "../persistence/schema/index.js";
