+++ assistant/src/approvals/scoped-approval-grants.ts
@@ -14,10 +14,10 @@
 import { and, eq, sql } from "drizzle-orm";
 import { v4 as uuid } from "uuid";
 
+import { rawChanges } from "../memory/raw-query.js";
 import { getDb } from "../persistence/db-connection.js";
 import { scopedApprovalGrants } from "../persistence/schema/index.js";
 import { getLogger } from "../util/logger.js";
-import { rawChanges } from "./raw-query.js";
 
 const log = getLogger("scoped-approval-grants");
