+++ assistant/src/calls/call-controller.ts
@@ -8,6 +8,7 @@
  * barge-in, state machine, guardian verification).
  */
 
+import { revokeScopedApprovalGrantsForContext } from "../approvals/scoped-approval-grants.js";
 import { loadConfig } from "../config/loader.js";
 import {
   expireCanonicalGuardianRequest,
@@ -18,7 +19,6 @@ import {
 import type { ServerMessage } from "../daemon/message-protocol.js";
 import type { TrustContext } from "../daemon/trust-context.js";
 import { getPublicBaseUrl } from "../inbound/public-ingress-urls.js";
-import { revokeScopedApprovalGrantsForContext } from "../memory/scoped-approval-grants.js";
 import { DAEMON_INTERNAL_ASSISTANT_ID } from "../runtime/assistant-scope.js";
 import { computeToolApprovalDigest } from "../security/tool-approval-digest.js";
 import { getCatalogProvider } from "../tts/provider-catalog.js";
