+++ assistant/src/calls/call-domain.ts
@@ -5,6 +5,7 @@
  * to these functions so business logic lives in one place.
  */
 
+import { revokeScopedApprovalGrantsForContext } from "../approvals/scoped-approval-grants.js";
 import { loadConfig } from "../config/loader.js";
 import { VALID_CALLER_IDENTITY_MODES } from "../config/schema.js";
 import type { AssistantConfig } from "../config/types.js";
@@ -15,7 +16,6 @@ import {
 } from "../inbound/public-ingress-urls.js";
 import { getOrCreateConversation } from "../memory/conversation-key-store.js";
 import { queueGenerateConversationTitle } from "../memory/conversation-title-service.js";
-import { revokeScopedApprovalGrantsForContext } from "../memory/scoped-approval-grants.js";
 import { getConversation } from "../persistence/conversation-crud.js";
 import { upsertBinding } from "../persistence/external-conversation-store.js";
 import { DAEMON_INTERNAL_ASSISTANT_ID } from "../runtime/assistant-scope.js";
