+++ assistant/src/calls/relay-server.ts
@@ -11,6 +11,7 @@ import { randomInt } from "node:crypto";
 import type { AdmissionPolicy, TrustVerdict } from "@vellumai/gateway-client";
 import type { ServerWebSocket } from "bun";
 
+import { revokeScopedApprovalGrantsForContext } from "../approvals/scoped-approval-grants.js";
 import { getCanonicalGuardianRequest } from "../contacts/canonical-guardian-store.js";
 import {
   getGuardianDelivery,
@@ -20,7 +21,6 @@ import {
 import { getAssistantName } from "../daemon/identity-helpers.js";
 import type { ServerMessage } from "../daemon/message-protocol.js";
 import type { TrustContext } from "../daemon/trust-context.js";
-import { revokeScopedApprovalGrantsForContext } from "../memory/scoped-approval-grants.js";
 import { addMessage } from "../persistence/conversation-crud.js";
 import { resolveGuardianName } from "../prompts/user-reference.js";
 import { notifyGuardianOfAccessRequest } from "../runtime/access-request-helper.js";
